# apiDoc/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /purchases:
    post:
      tags:
      - Purchase
      summary: Purchase a dish from a restaurant
      operationId: purchaseDish
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseHistory'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /transactions/top-users:
    get:
      tags:
      - Purchase
      summary: Find top x users by total transaction amount within a date range
      operationId: findTopTxUsers
      parameters:
      - name: top
        in: query
        description: "The specified top x (min = 1, max = 1000)"
        required: false
        schema:
          maximum: 1000
          minimum: 1
          type: integer
          format: int32
          default: 10
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTxAmount'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /transactions/sum:
    get:
      tags:
      - Purchase
      summary: Find total number and dollar value of transactions that happened within
        a date range
      operationId: findTxNumbAmount
      parameters:
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TxNumbAmount'
  /transactions/max-restaurants:
    get:
      tags:
      - Purchase
      summary: "Find most popular restaurants by transaction volume, either by number\
        \ of transactions or transaction dollar value"
      operationId: findMaxTxRestaurants
      parameters:
      - name: byAmount
        in: query
        description: To indicate it's calculated by transaction amount or number of
          transactions
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantTxAmount'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /transactions/user-count:
    get:
      tags:
      - Purchase
      summary: Find total number of users who made transactions above or below $v
        within a date range
      operationId: getUserCount
      parameters:
      - name: amount
        in: query
        description: The specified transaction amount
        required: true
        schema:
          minimum: 0
          type: number
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      - name: lessThan
        in: query
        description: To indicate less or more than the specified transaction amount
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Count'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /menus:
    put:
      tags:
      - Menu
      summary: "Edit dish name, dish price of menu"
      operationId: updateMenu
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMenu'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /menus/findByPrice:
    get:
      tags:
      - Menu
      summary: "List all dishes that are within a price range, sorted by price or\
        \ alphabetically"
      operationId: findMenusWithinPrices
      parameters:
      - name: maxPrice
        in: query
        description: Max dish price (min = 1.0)
        required: true
        schema:
          minimum: 1
          type: number
      - name: minPrice
        in: query
        description: Min dish price
        required: false
        schema:
          type: number
          default: 0
      - name: sortByPrice
        in: query
        description: Sorted by price or dish name
        required: false
        schema:
          type: boolean
          default: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /menus/findByDishName:
    get:
      tags:
      - Menu
      summary: Search dishes by dish name
      operationId: findMenusByDishName
      parameters:
      - name: dishName
        in: query
        description: The specified dish name
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  schemas:
    Purchase:
      required:
      - menuId
      type: object
      properties:
        userId:
          type: integer
          format: int64
        menuId:
          type: integer
          format: int64
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        msg:
          type: string
    PurchaseHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        menuId:
          type: integer
          format: int64
        dishName:
          type: string
        transactionAmount:
          type: number
        transactionDate:
          type: string
          format: date-time
    UserTxAmount:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        userName:
          type: string
        txAmount:
          type: number
    TxNumbAmount:
      type: object
      properties:
        txAmount:
          type: number
        txNumb:
          type: integer
          format: int64
    RestaurantTxAmount:
      type: object
      properties:
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        txAmount:
          type: number
        txNumb:
          type: integer
          format: int64
    Count:
      type: object
      properties:
        numb:
          type: integer
          format: int64
    EditMenu:
      required:
      - dishName
      type: object
      properties:
        menuId:
          type: integer
          format: int64
        dishName:
          maxLength: 500
          minLength: 0
          type: string
        price:
          minimum: 0.0
          exclusiveMinimum: false
          type: number
    Menu:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        dishName:
          type: string
        price:
          type: number
...