# src/main/java/idv/fd/purchase/dto/Count.java
package idv.fd.purchase.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Count {

    private Long numb;

}