# src/main/java/idv/fd/purchase/dto/TxNumbAmount.java
package idv.fd.purchase.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class TxNumbAmount {

    private Long txNumb;

    private BigDecimal txAmount;

}