# src/main/java/idv/fd/purchase/dto/UserTxAmount.java
package idv.fd.purchase.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class UserTxAmount {

    private Long userId;

    private String userName;

    private BigDecimal txAmount;

}